test: AuctionLotsForYou
