+++ src/schema/v2/homeView/__tests__/HomeViewSection.test.ts
@@ -155,54 +155,6 @@ describe("HomeViewSection", () => {
     })
   })
 
-  describe("AuctionLotsForYou", () => {
-    it("returns correct data", async () => {
-      const query = gql`
-        {
-          homeView {
-            section(id: "home-view-section-auction-lots-for-you") {
-              __typename
-              component {
-                title
-                behaviors {
-                  viewAll {
-                    buttonText
-                    href
-                    ownerType
-                  }
-                }
-              }
-              ownerType
-            }
-          }
-        }
-      `
-
-      const context = {
-        accessToken: "424242",
-      }
-
-      const { homeView } = await runQuery(query, context)
-
-      expect(homeView.section).toMatchInlineSnapshot(`
-        Object {
-          "__typename": "HomeViewSectionArtworks",
-          "component": Object {
-            "behaviors": Object {
-              "viewAll": Object {
-                "buttonText": "Browse All Artworks",
-                "href": null,
-                "ownerType": null,
-              },
-            },
-            "title": "Auction Lots for You",
-          },
-          "ownerType": "lotsByArtistsYouFollow",
-        }
-      `)
-    })
-  })
-
   describe("RecommendedArtworks", () => {
     it("returns lists of artworksConnection", async () => {
       const query = gql`
